Navigation Logic
The logic is essentially a manual image carousel.

Created a useEffect statement that listens for the useState 'element'. When the element useState is updated the useEffect scrolls into view the value of element.

The jsx has two buttons: scroll left and scroll right. Each button has an onclick evenetlistener that calls the function 'scrollTo.

The scrollTo function takes an arguement of true or false. The function creates a variable to be equal to the index of the elementsArr based on the 'elements' useState.